fix: add Account Selection to Account Setup Selection Screen.
