+++ app/src/bcsc-theme/features/account-transfer/AccountSetupSelectionScreen.tsx
@@ -1,17 +1,18 @@
+import { CardButton } from '@/bcsc-theme/components/CardButton'
 import GenericCardImage from '@/bcsc-theme/components/GenericCardImage'
 import { BCSCScreens, BCSCVerifyIdentityStackParams } from '@/bcsc-theme/types/navigators'
-import { BCState } from '@/store'
+import { BCDispatchAction, BCState } from '@/store'
 import { Button, ButtonType, ThemedText, useStore, useTheme } from '@bifold/core'
 import { useNavigation } from '@react-navigation/native'
 import { StackNavigationProp } from '@react-navigation/stack'
 
-import React from 'react'
+import React, { useCallback } from 'react'
 import { useTranslation } from 'react-i18next'
 import { StyleSheet, View } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 
 const AccountSetupSelectionScreen: React.FC = () => {
-  const [store] = useStore<BCState>()
+  const [store, dispatch] = useStore<BCState>()
   const { t } = useTranslation()
   const navigation = useNavigation<StackNavigationProp<BCSCVerifyIdentityStackParams>>()
   const { Spacing } = useTheme()
@@ -24,16 +25,22 @@ const AccountSetupSelectionScreen: React.FC = () => {
       justifyContent: 'space-between',
     },
     contentContainer: {
-      flex: 1,
+      flexGrow: 2,
       alignItems: 'center',
       justifyContent: 'center',
     },
     controlsContainer: {
-      marginTop: 'auto',
       width: '100%',
       gap: Spacing.sm,
     },
   })
+
+  const handleAccountSelect = useCallback(
+    (nickname: string) => {
+      dispatch({ type: BCDispatchAction.SELECT_ACCOUNT, payload: [nickname] })
+    },
+    [dispatch]
+  )
 
   return (
     <SafeAreaView edges={['bottom', 'left', 'right']} style={styles.container}>
@@ -42,26 +49,40 @@ const AccountSetupSelectionScreen: React.FC = () => {
         <ThemedText variant={'headerTitle'}>{t('Unified.AccountSetup.Title')}</ThemedText>
       </View>
 
-      <View style={styles.controlsContainer}>
-        <Button
-          buttonType={ButtonType.Primary}
-          title={t('Unified.AccountSetup.AddAccount')}
-          onPress={() => {
-            if (store.bcsc.completedNewSetup) {
-              navigation.navigate(BCSCScreens.SetupSteps)
-            } else {
-              navigation.navigate(BCSCScreens.NewSetup)
-            }
-          }}
-        />
-        <Button
-          buttonType={ButtonType.Secondary}
-          title={t('Unified.AccountSetup.TransferAccount')}
-          onPress={() => {
-            navigation.navigate(BCSCScreens.TransferAccountInformation)
-          }}
-        />
-      </View>
+      {store.bcsc.nicknames.length > 0 ? (
+        <View style={{ ...styles.controlsContainer, flexGrow: 1 }}>
+          <View style={{ marginBottom: Spacing.md }}>
+            <ThemedText variant={'headingFour'}>{t('Unified.AccountSelector.ContinueAs')}</ThemedText>
+          </View>
+
+          <View style={{ gap: Spacing.sm }}>
+            {Array.from(store.bcsc.nicknames).map((nickname) => (
+              <CardButton key={nickname} title={nickname} onPress={() => handleAccountSelect(nickname)} />
+            ))}
+          </View>
+        </View>
+      ) : (
+        <View style={styles.controlsContainer}>
+          <Button
+            buttonType={ButtonType.Primary}
+            title={t('Unified.AccountSetup.AddAccount')}
+            onPress={() => {
+              if (store.bcsc.completedNewSetup) {
+                navigation.navigate(BCSCScreens.SetupSteps)
+              } else {
+                navigation.navigate(BCSCScreens.NewSetup)
+              }
+            }}
+          />
+          <Button
+            buttonType={ButtonType.Secondary}
+            title={t('Unified.AccountSetup.TransferAccount')}
+            onPress={() => {
+              navigation.navigate(BCSCScreens.TransferAccountInformation)
+            }}
+          />
+        </View>
+      )}
     </SafeAreaView>
   )
 }
